Added files via upload
Me and my team used this data structure last year for siwir1 . I made relevant changes to it to initialize the boundary conditions.  Have a look at it and see if we can use this structure. I am trying to understand your code and thinking how we can do the smoothing.